# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "AskDelta": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "const": "delta",
            "default": "delta",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AskDelta",
        "type": "object"
      },
      "AskDone": {
        "properties": {
          "answer": {
            "$ref": "#/components/schemas/MessageOut"
          },
          "type": {
            "const": "done",
            "default": "done",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "answer"
        ],
        "title": "AskDone",
        "type": "object"
      },
      "AskError": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MessageOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "type": {
            "const": "error",
            "default": "error",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "answer"
        ],
        "title": "AskError",
        "type": "object"
      },
      "AskEvent": {
        "discriminator": {
          "mapping": {
            "delta": "#/components/schemas/AskDelta",
            "done": "#/components/schemas/AskDone",
            "error": "#/components/schemas/AskError",
            "sources": "#/components/schemas/AskSources",
            "start": "#/components/schemas/AskStart"
          },
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/AskStart"
          },
          {
            "$ref": "#/components/schemas/AskSources"
          },
          {
            "$ref": "#/components/schemas/AskDelta"
          },
          {
            "$ref": "#/components/schemas/AskDone"
          },
          {
            "$ref": "#/components/schemas/AskError"
          }
        ],
        "title": "AskEvent"
      },
      "AskRequest": {
        "properties": {
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Question",
            "type": "string"
          },
          "thread_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
...
      "FeedbackIn": {
        "properties": {
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "rating": {
            "$ref": "#/components/schemas/Rating"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "message_id",
          "rating"
        ],
        "title": "FeedbackIn",
        "type": "object"
      },
      "FeedbackOut": {
        "properties": {
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "rating": {
            "$ref": "#/components/schemas/Rating"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "rating",
          "reason",
          "updated_at"
        ],
        "title": "FeedbackOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Rating": {
        "enum": [
          "up",
          "down"
        ],
        "title": "Rating",
        "type": "string"
      },
...
      "ThreadDetail": {
        "properties": {
          "course_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lecture_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lecture Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "title": "Messages",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "lecture_id",
          "course_id",
          "title",
          "created_at",
          "updated_at",
          "messages"
        ],
        "title": "ThreadDetail",
        "type": "object"
      },
      "ThreadOut": {
        "description": "A conversation about a lecture or, with `course_id` set instead, a whole course.",
        "properties": {
          "course_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lecture_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lecture Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "lecture_id",
          "course_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ThreadOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Lecture Summariser API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/courses/{course_id}/ask": {
      "post": {
        "description": "Answer a question from every processed lecture in the course the caller may read.\nCitations name the lecture, like [L2 12:34]; each source's `lecture_label` says which\nlecture is L2.",
        "operationId": "ask_course_v1_courses__course_id__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskEvent"
                }
              }
            },
            "description": "Server-sent events. Each `data:` line is an AskEvent as JSON: start, sources, the answer in deltas, then done (or error)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask Course",
        "tags": [
          "qa"
        ]
      }
    },
    "/v1/courses/{course_id}/threads": {
      "get": {
        "description": "The caller's conversations about the course.",
        "operationId": "list_course_threads_v1_courses__course_id__threads_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThreadOut"
                  },
                  "title": "Response List Course Threads V1 Courses  Course Id  Threads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Course Threads",
        "tags": [
          "qa"
        ]
      }
    },
    "/v1/feedback": {
      "post": {
        "description": "Rate one of the caller's answers. Rating it again replaces the earlier rating.",
        "operationId": "feedback_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feedback",
        "tags": [
          "qa"
        ]
      }
    },
...
    "/v1/lectures/{lecture_id}/ask": {
      "post": {
        "description": "Answer a question from the lecture, citing it as [mm:ss]. The question and answer are\nsaved in a thread; send its `thread_id` to ask a follow-up. An answer the client\ndisconnects from isn't saved. Needs sign-in, within the user's quota of questions.",
        "operationId": "ask_v1_lectures__lecture_id__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskEvent"
                }
              }
            },
            "description": "Server-sent events. Each `data:` line is an AskEvent as JSON: start, sources, the answer in deltas, then done (or error)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "qa"
        ]
      }
    },
...
    "/v1/lectures/{lecture_id}/threads": {
      "get": {
        "description": "The caller's conversations about the lecture.",
        "operationId": "list_threads_v1_lectures__lecture_id__threads_get",
        "parameters": [
          {
            "in": "path",
            "name": "lecture_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lecture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThreadOut"
                  },
                  "title": "Response List Threads V1 Lectures  Lecture Id  Threads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Threads",
        "tags": [
          "qa"
        ]
      }
    },
...
    "/v1/threads/{thread_id}": {
      "delete": {
        "operationId": "delete_thread_v1_threads__thread_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Thread",
        "tags": [
          "qa"
        ]
      },
      "get": {
        "operationId": "get_thread_v1_threads__thread_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Thread",
        "tags": [
          "qa"
        ]
      }
    }
  }
}